pyproject.toml
[project]
name = "travel-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.86.0",
    "apscheduler>=3.11.2",
    "discord-py>=2.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]
